+++ packages/falcon-client/src/middlewares/routes/appHtmlMiddleware.js
@@ -8,7 +8,7 @@ import Html from '../../components/Html';
  * @param {{webpackAssets: object}} params webpack assets
  * @return {function(ctx: object, next: function): Promise<void>} Koa middleware
  */
-export default ({ webpackAssets }) => async ctx => {
+export default ({ webpackAssets, config }) => async ctx => {
   const { AppMarkup, client, chunkExtractor, helmetContext, serverTiming } = ctx.state;
   const renderTimer = serverTiming.start('HTML renderToString()');
 
@@ -19,6 +19,7 @@ export default ({ webpackAssets }) => async ctx => {
       helmetContext={helmetContext}
       state={client.extract()}
       i18nextState={extractI18nextState(ctx)}
+      googleTagManager={config.googleTagManager}
     >
       {AppMarkup}
     </Html>